Add extra flags to Makefile and eliminate warnings

+++ task.c
@@ -4,7 +4,7 @@
 #include "task.h"
 
 /* https://codereview.stackexchange.com/questions/29198/random-string-generator-in-c */
-char* form_rand_string(char* str, size_t size)
+char* form_rand_string(char* str, unsigned int size)
 {
     const char charset[] = "abcdefghijklmnopqrstuvwxyz";
     int charset_size = 25;
@@ -12,7 +12,7 @@ char* form_rand_string(char* str, size_t size)
     if (size > 0) 
     {
         --size;
-        for (size_t n = 0; n < size; n++) 
+        for (unsigned int n = 0; n < size; n++) 
         {
             key = rand() % charset_size;
             str[n] = charset[key];
@@ -25,12 +25,12 @@ char* form_rand_string(char* str, size_t size)
 long hash()
 {
     long hash = 0;
-    size_t str_size = 10;
+    unsigned int str_size = 10;
     char rand_str[str_size]; 
     form_rand_string(rand_str, str_size);
     const int a = rand() * 10;
     const int m = rand() * 15;
-    for (int i = 0; i < str_size; ++i)
+    for (unsigned int i = 0; i < str_size; ++i)
     {
         hash += (long)pow(a, str_size - (i+1)) * rand_str[i];
         hash = hash % m;
